-Dynamic artwork updating for Tunein, Yggdrasil Radio, and ANISON.FM, -Embedded artwork support for MP3, AAC, FLAC, OGG, WMA and AIFF, -Added automatic tiled rendering, -Allow vsync to work in 29.97 videos, -Options for setting music artwork in the screensaver, -Removed sample rate switching to avoid scratchy audio, -Added support for reading metadata from formats such as OGG and FLAC, -Specifying scan order in MKV files allows smooth playback in CRTs and HDTV deinterlacers, -Correct scale in 4:3 mode when using the Full Screen option, fixes 640x480 4:3 video from getting scaled, -Correct aspect ratio in widescreen mode, -Music art is now displayed as 188x horizontally, this size fixes a problem in the JPEG and PNG decoders where it would resample the image to a higher size and then downscale it resulting in noticeable blur, -Added a delay before showing the first few frames this avoids showing some garbage when loading from e.g. a channel forwarder, -Corrected a very strange problem when playing videos with subtitles, it would randomly freeze for different amounts of time on rare occasion, it created a large a/v desync, to fix it I enabled specific no-correctpts cases and kept the rest as it was, this fixed the random hangs and reduced the amount of problems introduced by using no-correctpts.
NOTE1: The tiled rendering implementation comes from mplayer-ce, done by Extrems.
NOTE2: The vsync fix for 30 fps videos comes from the PSF video mode implemented by Extrems.

+++ source/mplayer/libmpdemux/demux_mkv.c
@@ -2545,7 +2545,7 @@ static double fix_timestamp(demuxer_t *demuxer, mkv_track_t *track, double ts)
 		//find_prob++;
         // Assume that timestamps have been rounded to the timecode scale.
        double quant = MPMIN(mkv_d->tc_scale / 1e9, 0.001);
-       //double rts = rint_wii(ts / track->default_duration) * track->default_duration;
+   //    double rts = rint_wii(ts / track->default_duration) * track->default_duration;
        // rts = floor((ts / track->default_duration) + 0.5f) * track->default_duration;
        double rts = rint_wii(ts);// * track->default_duration;
         if (fabs(rts - ts) < quant)
@@ -2566,7 +2566,7 @@ static int handle_block(demuxer_t *demuxer, uint8_t *block, uint64_t length,
     uint32_t *lace_size;
     uint8_t laces, flags;
     int i, num, tmp, use_this_block = 1;
-    float current_pts;
+    double current_pts;
     int16_t time;
 
     /* first byte(s): track num */
